+++ modules/light-clients/08-wasm/CHANGELOG.md
@@ -43,7 +43,11 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ### API Breaking
 
+<<<<<<< HEAD
 * [\#6644](https://github.com/cosmos/ibc-go/pull/6644) api!: add `v2.MerklePath` for contract api `VerifyMembershipMsg` and `VerifyNonMembershipMsg` structs. Note, this requires a migration for existing client contracts to correctly handle deserialization of `MerklePath.KeyPath` which has changed from `repeated string` to `repeated bytes`. In JSON message structures this change is reflected as the `KeyPath` being a marshalled as a list of base64 encoded byte strings. 
+=======
+* [\#6937](https://github.com/cosmos/ibc-go/pull/6937) Remove `WasmConsensusHost` implementation of the `ConsensusHost` interface.
+>>>>>>> 2028e9a3 (imp!: removed 'ConsensusHost' interface (#6937))
 
 ### State Machine Breaking